# frontend/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "业务监控系统 API",
        "description": "业务监控系统后端接口文档，包含服务器资产管理、工作流编排、告警配置等功能模块。",
        "contact": {
            "name": "Monitor Team",
            "email": "monitor@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "开发环境"
        }
    ],
...
    "paths": {
...
        "/api/dashboard": {
            "get": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "获取所有仪表盘",
                "operationId": "list_4",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListDashboard"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "更新仪表盘",
                "operationId": "update_8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DashboardRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseDashboard"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "创建仪表盘",
                "operationId": "create_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DashboardRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseDashboard"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/widget": {
            "put": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "更新组件",
                "operationId": "updateWidget",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DashboardWidget"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseDashboardWidget"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/dashboard/{dashboardId}/widget": {
            "post": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "添加组件到仪表盘",
                "operationId": "addWidget",
                "parameters": [
                    {
                        "name": "dashboardId",
                        "in": "path",
                        "description": "仪表盘ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DashboardWidget"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseDashboardWidget"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/{dashboardId}/layout": {
            "post": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "批量保存仪表盘布局",
                "operationId": "saveLayout",
                "parameters": [
                    {
                        "name": "dashboardId",
                        "in": "path",
                        "description": "仪表盘ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/DashboardWidget"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/dashboard/{id}": {
            "get": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "获取仪表盘详情（含组件）",
                "operationId": "getById_6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "仪表盘ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMapStringObject"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "删除仪表盘",
                "operationId": "delete_8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "仪表盘ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/{dashboardId}/widgets": {
            "get": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "获取仪表盘的所有组件",
                "operationId": "getWidgets",
                "parameters": [
                    {
                        "name": "dashboardId",
                        "in": "path",
                        "description": "仪表盘ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListDashboardWidget"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/dashboard/widget/{id}": {
            "delete": {
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "删除组件",
                "operationId": "deleteWidget",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "组件ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "object",
                        "description": "响应数据"
                    }
                },
                "description": "统一API响应"
            },
...
            "DashboardRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID，更新时必填",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "仪表盘名称"
                    },
                    "description": {
                        "type": "string",
                        "description": "描述"
                    },
                    "gridCols": {
                        "type": "integer",
                        "description": "栅格列数",
                        "format": "int32"
                    },
                    "gridRows": {
                        "type": "integer",
                        "description": "栅格行数",
                        "format": "int32"
                    },
                    "cellHeight": {
                        "type": "integer",
                        "description": "单元格高度(px)",
                        "format": "int32"
                    },
                    "isDefault": {
                        "type": "integer",
                        "description": "是否为默认仪表盘",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "integer",
                        "description": "是否启用",
                        "format": "int32"
                    }
                },
                "description": "仪表盘请求"
            },
            "ApiResponseDashboard": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Dashboard"
                    }
                },
                "description": "统一API响应"
            },
            "Dashboard": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "仪表盘名称"
                    },
                    "gridCols": {
                        "type": "integer",
                        "description": "网格列数",
                        "format": "int32"
                    },
                    "gridRows": {
                        "type": "integer",
                        "description": "网格行数",
                        "format": "int32"
                    },
                    "cellHeight": {
                        "type": "integer",
                        "description": "单元格高度（像素）",
                        "format": "int32"
                    },
                    "widgetMargin": {
                        "type": "integer",
                        "description": "组件间距（像素）",
                        "format": "int32"
                    },
                    "backgroundConfig": {
                        "type": "string",
                        "description": "背景配置（JSON格式）"
                    },
                    "isDefault": {
                        "type": "integer",
                        "description": "是否为默认仪表盘（1=默认，0=非默认）",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "integer",
                        "description": "是否启用（1=启用，0=禁用）",
                        "format": "int32"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    }
                },
                "description": "仪表盘"
            },
            "DashboardWidget": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID",
                        "format": "int64"
                    },
                    "dashboardId": {
                        "type": "integer",
                        "description": "所属仪表盘ID",
                        "format": "int64"
                    },
                    "widgetType": {
                        "type": "string",
                        "description": "组件类型（TASK/WORKFLOW/CLOCK/TEXT/IMAGE）"
                    },
                    "sourceId": {
                        "type": "integer",
                        "description": "数据源ID（任务ID或工作流ID）",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string",
                        "description": "组件标题"
                    },
                    "gridX": {
                        "type": "integer",
                        "description": "网格X坐标",
                        "format": "int32"
                    },
                    "gridY": {
                        "type": "integer",
                        "description": "网格Y坐标",
                        "format": "int32"
                    },
                    "gridW": {
                        "type": "integer",
                        "description": "网格宽度（占用列数）",
                        "format": "int32"
                    },
                    "gridH": {
                        "type": "integer",
                        "description": "网格高度（占用行数）",
                        "format": "int32"
                    },
                    "displayConfig": {
                        "type": "string",
                        "description": "显示配置（JSON格式）"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "sourceResultType": {
                        "type": "string"
                    },
                    "sourceName": {
                        "type": "string"
                    },
                    "zindex": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "description": "仪表盘组件"
            },
            "ApiResponseDashboardWidget": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "$ref": "#/components/schemas/DashboardWidget"
                    }
                },
                "description": "统一API响应"
            },
...
            "ApiResponseMapStringObject": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object",
                            "description": "响应数据"
                        },
                        "description": "响应数据"
                    }
                },
                "description": "统一API响应"
            },
...
            "ApiResponseListDashboard": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "array",
                        "description": "响应数据",
                        "items": {
                            "$ref": "#/components/schemas/Dashboard"
                        }
                    }
                },
                "description": "统一API响应"
            },
            "ApiResponseListDashboardWidget": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "array",
                        "description": "响应数据",
                        "items": {
                            "$ref": "#/components/schemas/DashboardWidget"
                        }
                    }
                },
                "description": "统一API响应"
            },
...
        }
    }
}